requirements.txt
Django==1.7
dj-database-url==0.3.0
dj-static==0.0.6
django-filter==0.7
django-toolbelt==0.0.1
djangorestframework==2.3.14
gunicorn==19.1.0
psycopg2==2.5.3
static3==0.5.1
wsgiref==0.1.2
django-oauth-toolkit==0.7.2